Lazy embedding loading https://discord.com/channels/1014774730907209781/1014780368890630164/1122540657781776526
