+++ sdkit/models/model_loader/stable_diffusion/__init__.py
@@ -238,10 +238,9 @@ def load_diffusers_model(context: Context, model_path, config_file_path):
 
     gc(context)
 
-    if context.embeddings_path != None:
-        load_embeddings(context.embeddings_path, default_pipe)
-
     log.info("Loaded on diffusers")
+
+    context._loaded_embeddings = set(())
 
     return {
         "config": config,
@@ -251,60 +250,6 @@ def load_diffusers_model(context: Context, model_path, config_file_path):
         "inpainting": pipe_inpainting,
         "compel": compel,
     }
-
-
-def load_embeddings(path, default_pipe):
-    pt_files = list(path.rglob("*.pt"))
-    bin_files = list(path.rglob("*.bin"))
-    st_files = list(path.rglob("*.safetensors"))
-
-    for filename in pt_files + bin_files + st_files:
-        log.info(f"### Load embedding {filename} ###")
-        skip_embedding = False
-        embeds_name =  filename.name.split(".")[0]
-
-        embedding = load_tensor_file(filename)
-        dump_embedding_info(embedding)
-
-        model_dim = default_pipe.text_encoder.get_input_embeddings().weight.data[0].shape[0]
-
-        if "emb_params" in embedding.keys():
-            if model_dim != embedding["emb_params"].size(dim=-1):
-                skip_embedding = True
-        elif "<concept>" in embedding.keys():
-            if model_dim != embedding["<concept>"].size(dim=-1):
-                skip_embedding = True
-        elif "string_to_param" in embedding.keys():
-            for trained_token in embedding["string_to_param"]:
-                embeds = embedding["string_to_param"][trained_token]
-                if model_dim != embeds.size(dim=-1):
-                    skip_embedding = True
-                    continue
-        else:
-            log.info(f"Embedding {filename} has an unknown internal structure. Trying to load it anyways.")
-        
-        if skip_embedding:
-            log.info(f"Skipping embedding {filename}, due to incompatible embedding size, e.g. because this a StableDiffusion 2 embedding used with a StableDiffusion 1 model, or vice versa.")
-        else:
-            try:
-                default_pipe.load_textual_inversion(filename, embeds_name)
-            except:
-                log.error(f"Embedding {filename} can't be loaded. Proceeding without it!")
-                log.error(traceback.format_exc())
-
-
-def dump_embedding_info(embedding):
-    for key in dict(embedding).keys():
-        if key == "string_to_token":
-            for s in dict(embedding[key]).keys():
-                log.info(f"  - {key}: {s}")
-        elif key == "string_to_param":
-            for s in dict(embedding[key]).keys():
-                log.info(f"  - {key}: {s}")
-        elif key == "name" or key == "sd_checkpoint" or key == "sd_checkpoint_name" or key == "step":
-            log.info(f"  - {key}: '{embedding[key]}'")
-        else:
-            log.info(f"  # {key}")
 
 
 def test_and_fix_precision(context, model, config, attn_precision):
